# NewsPaper/news/views.py
import django_filters
from django.http import HttpResponseRedirect, HttpResponse
from django.shortcuts import render, get_object_or_404

# Create your views here.

# Импортируем класс, который говорит нам о том,
# что в этом представлении мы будем выводить список объектов из БД
from datetime import datetime

from django.contrib.auth.mixins import LoginRequiredMixin
from django.urls import reverse_lazy
from django.views import View
from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
from django.urls import reverse
from django.contrib.auth.models import User
from django.contrib.auth.decorators import login_required
from django.db.models import Exists, OuterRef
from django.shortcuts import render
from django.views.decorators.csrf import csrf_protect
from django.utils import timezone
from django.shortcuts import redirect
from django.utils.translation import gettext as _
# импортируем функцию для перевода
from django.utils.translation import activate, get_supported_language_variant #, LANGUAGE_SESSION_KEY
import pytz #  импортируем стандартный модуль для работы с часовыми поясами
from rest_framework.permissions import *
from rest_framework.renderers import JSONRenderer

from .models import Post, Author, Subscription, Category
from .filters import NewsFilter
from .forms import PostForm
from accounts.mixins import OwnerPermissionRequiredMixin
from django.core.cache import cache

from .templatetags.custom_tags import current_time

from rest_framework import viewsets, status
from rest_framework import permissions
from rest_framework.response import Response

from .serializers import *



class PostViewset(viewsets.ModelViewSet):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    permission_classes = [permissions.AllowAny]
    filter_backends = [django_filters.rest_framework.DjangoFilterBackend]
    filterset_fields = ['dateCreation', 'postCategory', 'title']
    # renderer_classes = [JSONRenderer]

    # def list(self, request, format=None ):
    #     return Response([])

    # def destroy(self, request, pk, format=None):
    #     instance = self.get_object()
    #     # instance.is_active = False
    #     instance.save()
    #     return Response(status=status.HTTP_204_NO_CONTENT)

    def get_queryset(self):
        queryset = Post.objects.all()
        dateCreation = self.request.query_params.get('dateCreation', None)
        postCategory = self.request.query_params.get('postCategory', None)
        title = self.request.query_params.get('title', None)
        if dateCreation is not None:
            queryset = queryset.filter(post__dateCreation=dateCreation)
        if postCategory is not None:
            queryset = queryset.filter(post__postCategory=postCategory)
        if title is not None:
            queryset = queryset.filter(post__title=title)
        return queryset

    # def get_permissions(self):
    #     """
    #     Instantiates and returns the list of permissions that this view requires.
    #     """
    #     if self.action == 'list':
    #         permission_classes = [AllowAny]
    #     if self.action == 'create':
    #         permission_classes = [IsAuthenticated]
    #     if self.action == 'update':
    #         permission_classes = [IsAuthenticated]
    #     if self.action == 'partial_update':
    #         permission_classes = [IsAuthenticated]
    #     if self.action == 'destroy':
    #         permission_classes = [IsAuthenticated]
    #     else:
    #         permission_classes = [IsAdminUser]
    #     return [permission() for permission in permission_classes]

# class AuthorViewset(viewsets.ModelViewSet):
#     queryset = Author.objects.all()
#     serializer_class = AuthorSerializer

class CategoryViewset(viewsets.ModelViewSet):
    renderer_classes = [JSONRenderer]
    queryset = Category.objects.all()
    serializer_class = CategorySerializer


# Create your views here.

# class Default(View):
#     def get(self, request):
#         current_time = timezone.now()
#
#         # .  Translators: This message appears on the home page only
#         # models = Post.objects.all()
#
#         context = {
#             # 'models': models,
#             'current_time': current_time,
#             'timezones': pytz.common_timezones #  добавляем в контекст все доступные часовые пояса
#         }
#         return HttpResponse(render(request, 'default.html', context))
#
#     #  по пост-запросу будем добавлять в сессию часовой пояс, который и будет обрабатываться написанным нами ранее middleware
#     def post(self, request):
#         request.session['django_timezone'] = request.POST['timezone']
#         return redirect('/')


# import logging
#
# logging = logging.getLogger(__name__)

class NewsList(ListView, View):
    model = Post
    ordering = '-dateCreation'
    template_name = 'news.html'
    context_object_name = 'news'
    paginate_by = 7

    # Переопределяем функцию получения списка товаров
    def get_queryset(self):
        # Получаем обычный запрос
        queryset = super().get_queryset()
        self.filterset = NewsFilter(self.request.GET, queryset)
        # Возвращаем из функции отфильтрованный список статей
        return self.filterset.qs

    def get_context_data(self, **kwargs):
        context = super().get_context_data(**kwargs)
        context['all_news'] = Post.objects.all()
        context['filterset'] = self.filterset
        context['current_time'] = timezone.localtime(timezone.now())
        context['timezones'] = pytz.common_timezones

        return context

    def get_success_url(self):
        return reverse('post_detail', kwargs={'pk': self.kwargs['pk']})

    def post(self, request):
        request.session['django_timezone'] = request.POST['timezone']
        return redirect('/news')


class NewsSearchList(ListView, View):
    # Указываем модель, объекты которой мы будем выводить
    model = Post
    # Поле, которое будет использоваться для сортировки объектов
    ordering = '-dateCreation'
    # Указываем имя шаблона, в котором будут все инструкции о том,
    # как именно пользователю должны быть показаны наши объекты
    # queryset = Post.objects.all()
    template_name = 'search.html'
    # Это имя списка, в котором будут лежать все объекты.
    # Его надо указать, чтобы обратиться к списку объектов в html-шаблоне.
    context_object_name = 'news'
    paginate_by = 7

    # Переопределяем функцию получения списка товаров
    def get_queryset(self):
        # Получаем обычный запрос
        queryset = super().get_queryset()
        # Используем наш класс фильтрации.
        # self.request.GET содержит объект QueryDict, который мы рассматривали
        # в этом юните ранее.
        # Сохраняем нашу фильтрацию в объекте класса,
        # чтобы потом добавить в контекст и использовать в шаблоне.
        self.filterset = NewsFilter(self.request.GET, queryset)
        # Возвращаем из функции отфильтрованный список статей
        return self.filterset.qs

    def get_context_data(self, **kwargs):
        context = super().get_context_data(**kwargs)
        context['all_news'] = Post.objects.all()
        context['filterset'] = self.filterset
        context['current_time'] = timezone.localtime(timezone.now())
        context['timezones'] = pytz.common_timezones
        return context

    def post(self, request):
        request.session['django_timezone'] = request.POST['timezone']
        return redirect('/search')

class PostDetail(DetailView):
    # Модель всё та же, но мы хотим получать информацию по отдельному посту
    model = Post
    # Используем другой шаблон — post.html
    template_name = 'post.html'
    # Название объекта, в котором будет выбранный пользователем пост
    context_object_name = 'post'

    def get_object(self, *args, **kwargs):  # переопределяем метод получения объекта, как ни странно
        obj = cache.get(f'product-{self.kwargs["pk"]}',
                        None)  # кэш очень похож на словарь, и метод get действует так же. Он забирает значение по ключу, если его нет, то забирает None.

        # если объекта нет в кэше, то получаем его и записываем в кэш
        if not obj:
            obj = super().get_object(queryset=self.queryset)
            cache.set(f'product-{self.kwargs["pk"]}', obj)

        return obj


class NewsCreate(LoginRequiredMixin, CreateView):
    permission_required = ('news.add_post',)
    raise_exception = True
    form_class = PostForm
    model = Post
    template_name = 'news/create.html'

    def form_valid(self, form):
        post = form.save(commit=False)
        post.categoryType = 'NW'
        form.instance.author = self.request.user.author
        return super().form_valid(form)

    def get_success_url(self):
        return reverse('news')


class ArticleCreate(LoginRequiredMixin, CreateView):
    permission_required = ('news.add_post',)
    raise_exception = True
    form_class = PostForm
    model = Post
    template_name = 'article/create.html'
    success_url = 'news'

    def form_valid(self, form):
        post = form.save(commit=False)
        post.categoryType = 'AR'
        form.instance.author = self.request.user.author
        return super().form_valid(form)

    def get_success_url(self):
        return reverse('news')


class NewsUpdate(OwnerPermissionRequiredMixin, UpdateView):
        permission_required = ('news.change_post',)
        raise_exception = True
        form_class = PostForm
        model = Post
        template_name = 'news/edit.html'

        def get_success_url(self):
            return reverse('post_detail',kwargs={'pk':self.kwargs['pk']})


class NewsDelete(OwnerPermissionRequiredMixin, DeleteView):
    permission_required = ('news.delete_post',)
    raise_exception = True
    model = Post
    template_name = 'news/delete.html'

    def get_success_url(self):
        return reverse('news')


class ArticleUpdate(OwnerPermissionRequiredMixin, UpdateView):
    permission_required = ('news.change_post',)
    raise_exception = True
    form_class = PostForm
    model = Post
    template_name = 'article/edit.html'

    def get_success_url(self):
        return reverse('post_detail', kwargs={'pk': self.kwargs['pk']})


class ArticleDelete(OwnerPermissionRequiredMixin, DeleteView):
    permission_required = ('news.change_post',)
    raise_exception = True
    model = Post
    template_name = 'article/delete.html'

    def get_success_url(self):
        return reverse('news')


# @login_required
# @csrf_protect
# def subscriptions(request):
#     if request.method == 'POST':
#         category_id = request.POST.get('category_id')
#         category = Category.objects.get(id=category_id)
#         action = request.POST.get('action')
#
#         if action == 'subscribe':
#             Subscription.objects.create(user=request.user, category=category)
#         elif action == 'unsubscribe':
#             Subscription.objects.filter(
#                 user=request.user,
#                 category=category,
#             ).delete()
#
#     categories_with_subscriptions = Category.objects.annotate(
#         user_subscribed=Exists(
#             Subscription.objects.filter(
#                 user=request.user,
#                 category=OuterRef('pk'),
#             )
#         )
#     ).order_by('name')
#     return render(
#         request,
#         'subscriptions.html',
#         {'categories': categories_with_subscriptions},
#     )

class CategoryListView(ListView):
    model = Post
    template_name = 'category_list.html'
    context_object_name = 'category_news_list'

    def get_queryset(self):
        self.category = get_object_or_404(Category, id=self.kwargs['pk'])
        queryset = Post.objects.filter(postCategory=self.category).order_by('-dateCreation')
        return queryset

    def get_context_data(self, **kwargs):
        context = super().get_context_data(**kwargs)
        context['is_not_subscriber'] = self.request.user not in self.category.subscribers.all()
        context['category'] = self.category
        return context


@login_required
@csrf_protect
def subscribe(request, pk):
    user = request.user
    category = Category.objects.get(id=pk)
    category.subscribers.add(user)

    message = 'Вы успешно подписались на рассылку новостей и статей категории'
    return render(request, 'subscribe.html', {'category': category, 'message': message})

@login_required
@csrf_protect
def unsubscribe(request, pk):
    user = request.user
    category = Category.objects.get(id=pk)
    category.subscribers.remove(user)

    message = 'Вы успешно отписались от рассылок новостей и статей категории'
    return render(request, 'unsubscribe.html', {'category': category, 'message': message})