# NewsPaper/static/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
  /news:
    get:
      operationId: listPosts
      description: ''
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: dateCreation
        required: false
        in: query
        description: dateCreation
        schema:
          type: string
      - name: postCategory
        required: false
        in: query
        description: postCategory
        schema:
          type: string
      - name: title
        required: false
        in: query
        description: title
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?offset=400&limit=100
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?offset=200&limit=100
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
          description: ''
      tags:
      - news
...
  /news/{id}:
    get:
      operationId: retrievePost
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this post.
        schema:
          type: string
      - name: dateCreation
        required: false
        in: query
        description: dateCreation
        schema:
          type: string
      - name: postCategory
        required: false
        in: query
        description: postCategory
        schema:
          type: string
      - name: title
        required: false
        in: query
        description: title
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
          description: ''
      tags:
      - news
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        categoryType:
          enum:
          - NW
          - AR
          type: string
        dateCreation:
          type: string
          format: date-time
          readOnly: true
        title:
          type: string
          maxLength: 128
        postCategory:
          type: array
          items:
            type: string
          readOnly: true
        text:
          type: string
      required:
      - title
      - text
...